Move VariantOverrides class to app/services
